requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.1
undetected-chromedriver>=3.5.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0

